Create Room Queue
I used a linked list to make a queue for the rooms. I am pretty sure that it will work, but I can't compile it without having access to the room class. Let me know if you see any problems or if you want me to add or change anything. Also, I changed the inventory I was working on to be a hash table that uses linked lists for its nodes to store items, rather than using a BST.