Fixed It. Almost Done With My Base Code.
Just need to edit one function bug left and it should be running smoothly. Then I can spend time on the actual monsters and floor descriptions and all that.

The function bug is that the answer string does indeed turn into the correct answer now, but the answer check still says it's wrong.